Backend: Secure login, registration, password encode, verification token generate, verification token email send, resend verification token if expired, password change token send by mail, change password

+++ CarService/src/main/java/com/EGEA1R/CarService/repository/FinanceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FinanceRepository extends JpaRepository<Finance, Integer> {
+public interface FinanceRepository extends JpaRepository<Finance, Long> {
 }
