+++ CarService/src/main/java/com/EGEA1R/CarService/repository/WorkdayRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkdayRepository extends JpaRepository<Workday, Integer> {
+public interface WorkdayRepository extends JpaRepository<Workday, Long> {
 }
